Refactor user help provider to allow proper extension for other documentation parts
